### 2024-07-20 #7016 @editorialbot
[PRE REVIEW]: Fundus Image Toolbox: A Python package for fundus image processing
**Submitting author:** @juliusge (<a href="http://orcid.org/0009-0007-7286-0017">Julius Gervelmeyer</a>)
**Repository:** https://github.com/berenslab/fundus_image_toolbox
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.0.1
