### 2024-07-22 #7016 @juliusge
From the list of people reviewing for JOSS, I think
- NMontanaBrown
- pchlap
- saeedahmadicp
could be a good fit as they mention Python, Medial Image Analysis and Deep Learning.
