### 2024-08-03 #7016 @osorensen
@NMontanaBrown, @pchlap, @saeedahmadicp, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
