### 2024-08-05 #7016 @saeedahmadicp
@osorensen Yes, I am ready to review the paper. Thanks for considering.
