### 2024-08-23 #7101 @osorensen
@ojasiiitd @sashakolpakov; after checking with some colleagues, it seems like GitHub does not provide contributor statistics for repositories where there is only a single contributor. This seems to be the case here, as we can see from the commit history: https://github.com/berenslab/fundus_image_toolbox/commits/main/
