### 2024-08-23 #7101 @juliusge
Hi there, I just noticed this. So I merged my private repositories after developing single models and then transferred the code to the current repository at my lab's github. Contribution stats would hence not be very meaningful, as most of the development was already done when this repo was created. I can offer to screenshot stats from some related private repos, if you need that. And, yes, I was the only contributor on github (though not in real life). Thank you all for your work and please let me know in case I should not interfere in this thread.
Best, Julius
