### 2024-10-08 #7101 @saeedahmadicp
Dear @osorensen ,

I have carefully reviewed the submission you sent. The article and the accompanying package aim to provide a toolkit for analyzing Fundus Images including quality prediction, fovea, and optic disc center localization. The transcript of the package is aligned with the journal guidelines and provides the summary, statement of needs, and details about the tools. However, a few issues need to be addressed in this version, which are:

- The documentation for the package is not sufficient and requires an extension to cover all the functionalities in detail and some more usecases and examples
- Upon several tests, the package provides the mentioned functionalities; however, there are a few issues related to path management that need to be addressed. More specifically, the package is unable to resolve the paths properly in the Windows system.

![image](https://github.com/user-attachments/assets/de88a883-f2b5-44a7-9456-481d8db56672)

![image](https://github.com/user-attachments/assets/67f2bbbc-b821-46b4-9318-e1eb1632a4ee)

- The package needs to be uploaded and managed on the PyPI repository for ease of use and easy integration with existing code, as the current installation method requires a separate conda environment.

- Also, the authors haven't provided any guidelines for contributions to the repository. Additionally, the contribution of other authors in the paper is not mentioned explicitly nor implicitly in the GitHub repository and installation from GitHub.

**In summary, this work provides a valuable toolkit for fundus image processing that can significantly reduce effort for the community. However, to maximize its utility, the authors should consider enhancing the documentation, providing clear contribution guidelines, and maintaining the package on PyPI. These improvements would make the tool more accessible, encourage community involvement, and simplify integration with existing workflows, thereby increasing its overall impact in the field.**
