### 2024-10-28 4bb10c4
use pathlib in fovea and quality modules
