+++ fundus_quality_prediction/fundus_quality_prediction/scripts/model.py
@@ -1,6 +1,7 @@
 from types import SimpleNamespace
 from typing import Union
 import yaml
+from pathlib import Path
 
 import matplotlib as mpl
 import matplotlib.pyplot as plt
@@ -15,81 +16,115 @@ from datetime import datetime
 import os
 from PIL import ImageDraw, Image
 import numpy as np
-from sklearn.metrics import accuracy_score, f1_score, precision_score, recall_score, auc, roc_curve, precision_recall_curve
+from sklearn.metrics import (
+    accuracy_score,
+    f1_score,
+    precision_score,
+    recall_score,
+    auc,
+    roc_curve,
+    precision_recall_curve,
+)
 from fundus_utilities import ImageTorchUtils as Img
 from fundus_utilities import seed_everything
 
 from .transforms import get_unnormalization, get_transforms
 from .default import ENSEMBLE_MODELS, MODELS_DIR
 
-def download_weights(url = "https://zenodo.org/records/11174749/files/weights.tar.gz"):
-    os.makedirs(MODELS_DIR, exist_ok=True)
-    print('Downloading weights...')
-    os.system(f'wget -q {url} -O {MODELS_DIR}/weights.tar.gz')
-    print('Extracting weights...')
-    os.system(f'tar -xzf {MODELS_DIR}/weights.tar.gz -C {MODELS_DIR}')
-    print('Removing tar file...')
-    os.system(f'rm {MODELS_DIR}/weights.tar.gz')
-    print('Done')
+# def download_weights(url = "https://zenodo.org/records/11174749/files/weights.tar.gz"):
+#     os.makedirs(MODELS_DIR, exist_ok=True)
+#     print('Downloading weights...')
+#     os.system(f'wget -q {url} -O {MODELS_DIR}/weights.tar.gz')
+#     print('Extracting weights...')
+#     os.system(f'tar -xzf {MODELS_DIR}/weights.tar.gz -C {MODELS_DIR}')
+#     print('Removing tar file...')
+#     os.system(f'rm {MODELS_DIR}/weights.tar.gz')
+#     print('Done')
 
-def plot_quality(fundus, conf:float, label:int, threshold:float=0.5):
+
+def download_weights(url="https://zenodo.org/records/11174749/files/weights.tar.gz"):
+    MODELS_DIR.mkdir(parents=True, exist_ok=True)
+    print("Downloading weights...")
+    os.system(f'wget -q {url} -O {MODELS_DIR / "weights.tar.gz"}')
+    print("Extracting weights...")
+    os.system(f'tar -xzf {MODELS_DIR / "weights.tar.gz"} -C {MODELS_DIR}')
+    print("Removing tar file...")
+    os.system(f'rm {MODELS_DIR / "weights.tar.gz"}')
+    print("Done")
+
+
+def plot_quality(fundus, conf: float, label: int, threshold: float = 0.5):
     """Plots the fundus image with a colorbar indicating the confidence score.
-    
+
     Args:
         fundus (str, np.ndarray): Fundus image path or image as numpy array
         conf (float): Confidence score
         label (int): Label (0: ungradable, 1: gradable)
         threshold (float, optional): Threshold for binary classification. Defaults to 0.5.
-    
+
     Returns:
         None
     """
     # Image to array with (H, W, C) shape
     fundus = Img(fundus).to_tensor().squeeze().to_numpy().set_channel_dim(-1).img
-    
+
     if Img(fundus).is_batch_like():
         raise ValueError("Pass a single image, not a batch.")
 
     fig = plt.figure(figsize=(4.3, 4))
[…]
