+++ fundus_fovea_od_localization/fundus_fovea_od_localization/transforms_multi.py
@@ -3,329 +3,380 @@ import torch
 import torchvision.transforms.functional as tf
 import numpy as np
 
+
 class Resize:
-  '''Resize the image and coordinate labels.'''
-  def __init__(self, new_size=(350, 350)):
-    self.new_width = new_size[0]
-    self.new_height = new_size[1]
+    """Resize the image and coordinate labels."""
 
-  def __call__(self, image_label_sample):
-    image = image_label_sample[0]
-    f_label, od_label = image_label_sample[1]
-    f_x, f_y = f_label
-    od_x, od_y = od_label
-    original_width, original_height = image.size
-    image_new = tf.resize(image, (self.new_height, self.new_width))
+    def __init__(self, new_size=(350, 350)):
+        self.new_width = new_size[0]
+        self.new_height = new_size[1]
 
-    f_x_new = f_x * self.new_width /original_width
-    f_y_new = f_y * self.new_height / original_height
-    f_label_new = (f_x_new, f_y_new)
+    def __call__(self, image_label_sample):
+        image = image_label_sample[0]
+        f_label, od_label = image_label_sample[1]
+        f_x, f_y = f_label
+        od_x, od_y = od_label
+        original_width, original_height = image.size
+        image_new = tf.resize(image, (self.new_height, self.new_width))
 
-    od_x_new = od_x * self.new_width /original_width
-    od_y_new = od_y * self.new_height / original_height
-    od_label_new = (od_x_new, od_y_new)
+        f_x_new = f_x * self.new_width / original_width
+        f_y_new = f_y * self.new_height / original_height
+        f_label_new = (f_x_new, f_y_new)
 
-    return image_new, (f_label_new, od_label_new)
-  
+        od_x_new = od_x * self.new_width / original_width
+        od_y_new = od_y * self.new_height / original_height
+        od_label_new = (od_x_new, od_y_new)
+
+        return image_new, (f_label_new, od_label_new)
+
+
 class ResizedCenterCrop:
-  '''Resize the image and coordinate labels.
+    """Resize the image and coordinate labels.
     First crops a square from the center, then resizes.
-  '''
-  
-  def __init__(self, new_size=(350, 350)):
-    self.new_width = new_size[0]
-    self.new_height = new_size[1]
+    """
 
-  def __call__(self, image_label_sample):
-    image = image_label_sample[0]
-    f_label, od_label = image_label_sample[1]
-    f_x, f_y = f_label
-    od_x, od_y = od_label
-    original_width, original_height = image.size
+    def __init__(self, new_size=(350, 350)):
+        self.new_width = new_size[0]
+        self.new_height = new_size[1]
 
-    image = tf.center_crop(image, original_height) # assume landscape
-    f_x, f_y = f_x - (original_width - original_height) / 2, f_y
-    od_x, od_y = od_x - (original_width - original_height) / 2, od_y
-    original_width = original_height
+    def __call__(self, image_label_sample):
+        image = image_label_sample[0]
+        f_label, od_label = image_label_sample[1]
+        f_x, f_y = f_label
+        od_x, od_y = od_label
+        original_width, original_height = image.size
 
-    image = tf.resize(image, (self.new_height, self.new_width))
-    f_x_new = f_x * self.new_width /original_width
-    f_y_new = f_y * self.new_height / original_height
-    f_label_new = (f_x_new, f_y_new)
+        image = tf.center_crop(image, original_height)  # assume landscape
+        f_x, f_y = f_x - (original_width - original_height) / 2, f_y
[…]
