### 2024-10-30 0537ac9
fixes to make usage file of fovea model work with Pathlib
