### 2024-10-30 82814d0
use pathlib for quality prediction code
