### 2024-11-22 7544304
Update model_multi.py for Windows Paths
