### 2024-11-22 3df0ff1
exchange wget through cross-platform function
