+++ fundus_registration/fundus_registration/inference.py
@@ -1,6 +1,7 @@
 # %%
 from typing import List, Union
 import os
+import requests
 import torch
 import cv2
 import numpy as np
@@ -12,30 +13,54 @@ import pandas as pd
 import pickle
 from fundus_utilities import ImageTorchUtils as Img
 
-from .SuperRetina import pre_processing, simple_nms, remove_borders, sample_keypoint_desc, SuperRetina
+from .SuperRetina import (
+    pre_processing,
+    simple_nms,
+    remove_borders,
+    sample_keypoint_desc,
+    SuperRetina,
+)
 
-WEIGHT_PATH = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'SuperRetina/save/SuperRetina.pth')
+WEIGHT_PATH = os.path.join(
+    os.path.dirname(os.path.abspath(__file__)), "SuperRetina/save/SuperRetina.pth"
+)
 IMG_SIZE = 512
-DEFAULT_CONFIG = {  
-                'device': 'cuda:0',
-                'use_matching_trick': True,
-                'nms_size': 1,
-                'nms_thresh': 0.0005,
-                'knn_thresh': 0.85
-                }
+DEFAULT_CONFIG = {
+    "device": "cuda:0",
+    "use_matching_trick": True,
+    "nms_size": 1,
+    "nms_thresh": 0.0005,
+    "knn_thresh": 0.85,
+}
 
-def download_weights(url = "https://zenodo.org/records/11241985/files/SuperRetina.pth"):
+
+def wget(link, target):
+    # platform independent wget alternative
+    headers = {
+        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/87.0.4280.141 Safari/537.36"
+    }
+    r = requests.get(link, headers=headers, stream=True)
+    downloaded_file = open(target, "wb")
+
+    for chunk in r.iter_content(chunk_size=8192):
+        if chunk:
+            downloaded_file.write(chunk)
+
+
+def download_weights(url="https://zenodo.org/records/11241985/files/SuperRetina.pth"):
     os.makedirs(os.path.dirname(WEIGHT_PATH), exist_ok=True)
-    print('Downloading weights...')
-    os.system(f'wget -q {url} -O {WEIGHT_PATH}')
-    print('Done')
+    print("Downloading weights...")
+    # os.system(f'wget -q {url} -O {WEIGHT_PATH}')
+    wget(url, WEIGHT_PATH)
+    print("Done")
 
-def get_mask(filename, masks: pd.DataFrame= None, show=True, binarize=True):
+
+def get_mask(filename, masks: pd.DataFrame = None, show=True, binarize=True):
     directory = os.path.dirname(filename)
     if masks is None:
-        p = os.path.join(directory, 'masks.pkl')
-        masks = pickle.load(open(p, 'rb'))
-    
+        p = os.path.join(directory, "masks.pkl")
+        masks = pickle.load(open(p, "rb"))
+
     row = masks.loc[masks["image"] == os.path.basename(filename)]
     mask = row["mask"].values[0]
     size = row["size"].values[0]
@@ -51,7 +76,7 @@ def get_mask(filename, masks: pd.DataFrame= None, show=True, binarize=True):
         plt.show()
         plt.close()
 
-    # assert mask not empty 
+    # assert mask not empty
     assert np.sum(mask) > 0
 
     return mask
@@ -70,6 +95,7 @@ def any_to_image_array(image):
         image = cv2.cvtColor(image, cv2.COLOR_RGB2BGR)
     return image
 
+
 def any_to_greyscale(image):
[…]
