### 2024-11-22 6a4cda9
remove old submodule structure from dependencies
