+++ fundus_registration/SuperRetina/predictor.py
@@ -5,8 +5,12 @@ import numpy as np
 import matplotlib.pyplot as plt
 from torchvision import transforms
 
-from .common.common_util import pre_processing, simple_nms, remove_borders, \
-    sample_keypoint_desc
+from .common.common_util import (
+    pre_processing,
+    simple_nms,
+    remove_borders,
+    sample_keypoint_desc,
+)
 from .model.super_retina import SuperRetina
 
 from PIL import Image
@@ -16,37 +20,38 @@ import os
 class Predictor:
     def __init__(self, config):
 
-        predict_config = config['PREDICT']
+        predict_config = config["PREDICT"]
 
-        device = predict_config['device']
+        device = predict_config["device"]
         device = torch.device(device if torch.cuda.is_available() else "cpu")
 
-        model_save_path = predict_config['model_save_path']
-        self.nms_size = predict_config['nms_size']
-        self.nms_thresh = predict_config['nms_thresh']
+        model_save_path = predict_config["model_save_path"]
+        self.nms_size = predict_config["nms_size"]
+        self.nms_thresh = predict_config["nms_thresh"]
         self.scale = 8
-        self.knn_thresh = predict_config['knn_thresh']
+        self.knn_thresh = predict_config["knn_thresh"]
 
         self.image_width = None
         self.image_height = None
 
-        self.model_image_width = predict_config['model_image_width']
-        self.model_image_height = predict_config['model_image_height']
+        self.model_image_width = predict_config["model_image_width"]
+        self.model_image_height = predict_config["model_image_height"]
 
-        checkpoint = torch.load(model_save_path, map_location=device)
+        checkpoint = torch.load(model_save_path, map_location=device, weights_only=True)
         model = SuperRetina()
-        model.load_state_dict(checkpoint['net'])
+        model.load_state_dict(checkpoint["net"])
         model.to(device)
         model.eval()
         self.device = device
         self.model = model
         self.knn_matcher = cv2.BFMatcher(cv2.NORM_L2)
 
-        self.trasformer = transforms.Compose([
-            transforms.Resize((self.model_image_height, self.model_image_width)),
-            transforms.ToTensor(),
-
-        ])
+        self.trasformer = transforms.Compose(
+            [
+                transforms.Resize((self.model_image_height, self.model_image_width)),
+                transforms.ToTensor(),
+            ]
+        )
 
     def image_read(self, query_path, refer_path, query_is_image=False):
         if query_is_image:
@@ -69,7 +74,9 @@ class Predictor:
 
         return query_image, refer_image
 
-    def draw_result(self, query_image, refer_image, cv_kpts_query, cv_kpts_refer, matches, status):
+    def draw_result(
+        self, query_image, refer_image, cv_kpts_query, cv_kpts_refer, matches, status
+    ):
         def drawMatches(imageA, imageB, kpsA, kpsB, matches, status):
             # initialize the output visualization image
             (hA, wA) = imageA.shape[:2]
@@ -99,12 +106,14 @@ class Predictor:
         query_np = np.array([kp.pt for kp in cv_kpts_query])
         refer_np = np.array([kp.pt for kp in cv_kpts_refer])
         refer_np[:, 0] += query_image.shape[1]
-        matched_image = drawMatches(query_image, refer_image, cv_kpts_query, cv_kpts_refer, matches, status)
+        matched_image = drawMatches(
+            query_image, refer_image, cv_kpts_query, cv_kpts_refer, matches, status
+        )
         plt.figure(dpi=300)
-        plt.scatter(query_np[:, 0], query_np[:, 1], s=1, c='r')
[…]

[…]
