+++ MANIFEST.in
@@ -1,4 +1,5 @@
 include Readme.md
 include requirements.txt
-include fundus1.jog
-include fundus2.jog
+include fundus1.jpg
+include fundus2.jpg
+include fundus_fovea_od_localization/data/ADAM+IDRID+REFUGE_df.csv
