+++ test/test_vessel_segmentation.py
@@ -1,13 +1,21 @@
-import os 
+import os
 import unittest
 import numpy as np
 import torch
 from PIL import Image
-from fundus_vessel_segmentation import load_segmentation_ensemble, ensemble_predict_segmentation, plot_masks, save_masks, load_masks_from_filenames, FR_UNet
+from fundus_image_toolbox.vessel_segmentation import (
+    load_segmentation_ensemble,
+    ensemble_predict_segmentation,
+    plot_masks,
+    save_masks,
+    load_masks_from_filenames,
+    FR_UNet,
+)
 
 DIR = os.path.join(os.path.dirname(__file__))
 
-fundus1_path = os.path.join(DIR, '..', 'fundus1.jpg')
+fundus1_path = os.path.join(DIR, "..", "fundus1.jpg")
+
 
 class TestVesselSegmentation(unittest.TestCase):
     def setUp(self):
@@ -23,18 +31,22 @@ class TestVesselSegmentation(unittest.TestCase):
 
     def test_ensemble_predict_segmentation(self):
         # Test the ensemble_predict_segmentation function
-        segmentation = ensemble_predict_segmentation(self.model, self.image, device=self.device)
+        segmentation = ensemble_predict_segmentation(
+            self.model, self.image, device=self.device
+        )
         self.assertIsInstance(segmentation, np.ndarray)
         self.assertEqual(segmentation.shape, (self.image.height, self.image.width))
 
     def test_plot_masks(self):
         # Test the plot_masks function
-        segmentation = ensemble_predict_segmentation(self.model, self.image, device=self.device)
+        segmentation = ensemble_predict_segmentation(
+            self.model, self.image, device=self.device
+        )
         plot_masks(self.image, segmentation)
 
     def test_save_and_load_masks(self):
         # Test the save_masks and load_masks_from_filenames functions
-        target_dir = os.path.join(DIR, 'masks')
+        target_dir = os.path.join(DIR, "masks")
         mask = ensemble_predict_segmentation(self.model, self.image, device=self.device)
         save_masks(fundus1_path, mask, target_dir)
         loaded_mask = load_masks_from_filenames(fundus1_path, target_dir)
@@ -45,5 +57,6 @@ class TestVesselSegmentation(unittest.TestCase):
         for file in os.listdir(os.path.join(target_dir)):
             os.remove(os.path.join(target_dir, file))
 
-if __name__ == '__main__':
+
+if __name__ == "__main__":
     unittest.main()
