+++ test/test_utilities_image_torch_utils.py
@@ -6,16 +6,17 @@ from PIL import Image
 import cv2
 import torch
 from torchvision.transforms.functional import to_tensor
-from fundus_utilities import ImageTorchUtils as Img
+from fundus_image_toolbox.utilities import ImageTorchUtils as Img
 
 DIR = os.path.join(os.path.dirname(__file__))
 
-fundus1_path = os.path.join(DIR, '..', 'fundus1.jpg')
-fundus2_path = os.path.join(DIR, '..', 'fundus2.jpg')
+fundus1_path = os.path.join(DIR, "..", "fundus1.jpg")
+fundus2_path = os.path.join(DIR, "..", "fundus2.jpg")
 
 fundus1_plt = plt.imread(fundus1_path)
 fundus1_img = Image.open(fundus1_path)
 fundus1_cv2 = cv2.imread(fundus1_path)
+
 
 class TestImageTorchUtils(unittest.TestCase):
     def test_to_tensor(self):
@@ -26,9 +27,14 @@ class TestImageTorchUtils(unittest.TestCase):
         outs = [Img(img).to_tensor().img for img in imgs_rgb]
 
         # Convert to rgb tensor from bgr image
-        bgr_out = Img(fundus1_cv2).to_tensor(from_cspace='bgr').img
+        bgr_out = Img(fundus1_cv2).to_tensor(from_cspace="bgr").img
         # Alternatively, use a sequence of calls:
-        bgr_out = Img(fundus1_cv2).to_tensor().to_cspace(from_cspace='bgr', to_cspace='rgb').img
+        bgr_out = (
+            Img(fundus1_cv2)
+            .to_tensor()
+            .to_cspace(from_cspace="bgr", to_cspace="rgb")
+            .img
+        )
         outs.append(bgr_out)
 
         assert all([torch.equal(outs[0], out) for out in outs[1:]])
@@ -39,16 +45,25 @@ class TestImageTorchUtils(unittest.TestCase):
         img_bgr = fundus1_cv2
 
         # Convert from rgb to bgr
-        outs = [Img(img).to_tensor().to_cspace(from_cspace='rgb', to_cspace='bgr').img for img in imgs_rgb]
+        outs = [
+            Img(img).to_tensor().to_cspace(from_cspace="rgb", to_cspace="bgr").img
+            for img in imgs_rgb
+        ]
 
         assert all([torch.equal(outs[0], out) for out in outs[1:]])
         assert torch.equal(Img(img_bgr).to_tensor().img, outs[0])
 
         # Convert to grey scale
-        outs = [Img(img).to_tensor().to_cspace(from_cspace='rgb', to_cspace='grey').img for img in imgs_rgb]
+        outs = [
+            Img(img).to_tensor().to_cspace(from_cspace="rgb", to_cspace="grey").img
+            for img in imgs_rgb
+        ]
 
         assert all([torch.equal(outs[0], out) for out in outs[1:]])
-        assert torch.equal(Img(img_bgr).to_tensor().to_cspace(from_cspace='bgr', to_cspace='grey').img, outs[0])
+        assert torch.equal(
+            Img(img_bgr).to_tensor().to_cspace(from_cspace="bgr", to_cspace="grey").img,
+            outs[0],
+        )
 
     def test_set_channel_dim(self):
         # Test set_channel_dim function
@@ -59,7 +74,7 @@ class TestImageTorchUtils(unittest.TestCase):
 
         # Check if the channel dimension is the last dimension
         assert all([out.shape[-1] == 3 for out in outs])
-    
+
     def test_to_pil(self):
         # Test to_pil function
         imgs = [fundus1_path, fundus1_plt, fundus1_img, fundus1_cv2]
@@ -117,14 +132,24 @@ class TestImageTorchUtils(unittest.TestCase):
[…]
