+++ quality_prediction/scripts/dataloader.py
@@ -5,7 +5,7 @@ import pandas as pd
 import torch
 
 from .dataset import FundusQualityDataset
-from fundus_utilities import ImbalancedDatasetSampler
+from fundus_image_toolbox.utilities import ImbalancedDatasetSampler
 from .default import DATA_ROOT
 
 

[…]
