+++ MANIFEST.in
@@ -3,3 +3,5 @@ include requirements.txt
 include fundus1.jpg
 include fundus2.jpg
 include od_localization/data/ADAM+IDRID+REFUGE_df.csv
+include 0_example_usage/*
+include 1_read_more/*

[…]
