### 2024-11-22 fdff7ef
fix path management and folder structure
