### 2024-11-22 f93afbf
fix path to image
