+++ fundus_image_toolbox/vessel_segmentation/segmentation/utils/model_definition.py
@@ -1,213 +1,0 @@
-import torch
-import torch.nn as nn
-from timm.models.layers import trunc_normal_
-
-
-class InitWeights_He(object):
-    def __init__(self, neg_slope=1e-2):
-        self.neg_slope = neg_slope
-
-    def __call__(self, module):
-        if isinstance(module, nn.Conv3d) or isinstance(module, nn.Conv2d) or isinstance(module, nn.ConvTranspose2d) or isinstance(module, nn.ConvTranspose3d):
-            module.weight = nn.init.kaiming_normal_(module.weight, a=self.neg_slope)
-            if module.bias is not None:
-                module.bias = nn.init.constant_(module.bias, 0)
-        elif isinstance(module, nn.Linear):
-            trunc_normal_(module.weight, std=self.neg_slope)
-            if isinstance(module, nn.Linear) and module.bias is not None:
-                nn.init.constant_(module.bias, 0)
-        elif isinstance(module, nn.LayerNorm):
-            nn.init.constant_(module.bias, 0)
-            nn.init.constant_(module.weight, 1.0)
-
-
-class conv(nn.Module):
-    def __init__(self, in_c, out_c, dp=0.0):
-        super(conv, self).__init__()
-        self.in_c = in_c
-        self.out_c = out_c
-        self.dp = dp
-        self.conv = nn.Sequential(
-            nn.Conv2d(out_c, out_c, kernel_size=3, padding=1, bias=False),
-            nn.BatchNorm2d(out_c),
-            nn.Dropout2d(self.dp ),
-            nn.LeakyReLU(0.1, inplace=True),
-            nn.Conv2d(out_c, out_c, kernel_size=3, padding=1, bias=False),
-            nn.BatchNorm2d(out_c),
-            nn.Dropout2d(self.dp),
-            nn.LeakyReLU(0.1, inplace=True))
-        self.relu = nn.LeakyReLU(0.1, inplace=True)
-
-    def forward(self, x):
-        res = x
-        x = self.conv(x)
-        out = x + res
-        out = self.relu(out)
-        return x
-
-
-class feature_fuse(nn.Module):
-    def __init__(self, in_c, out_c):
-        super(feature_fuse, self).__init__()
-        self.conv11 = nn.Conv2d(
-            in_c, out_c, kernel_size=1, padding=0, bias=False)
-        self.conv33 = nn.Conv2d(
-            in_c, out_c, kernel_size=3, padding=1, bias=False)
-        self.conv33_di = nn.Conv2d(
-            in_c, out_c, kernel_size=3, padding=2, bias=False, dilation=2)
-        self.norm = nn.BatchNorm2d(out_c)
-
-    def forward(self, x):
-        x1 = self.conv11(x)
-        x2 = self.conv33(x)
-        x3 = self.conv33_di(x)
-        out = self.norm(x1+x2+x3)
-        return out
-
-
-class up(nn.Module):
-    def __init__(self, in_c, out_c, dp=0):
-        super(up, self).__init__()
-        self.up = nn.Sequential(
-            nn.ConvTranspose2d(in_c, out_c, kernel_size=2,
-                               padding=0, stride=2, bias=False),
-            nn.BatchNorm2d(out_c),
-            nn.LeakyReLU(0.1, inplace=False))
-
-    def forward(self, x):
-        x = self.up(x)
-        return x
-
-
-class down(nn.Module):
-    def __init__(self, in_c, out_c, dp=0):
-        super(down, self).__init__()
-        self.down = nn.Sequential(
-            nn.Conv2d(in_c, out_c, kernel_size=2,
-                      padding=0, stride=2, bias=False),
-            nn.BatchNorm2d(out_c),
[…]
