+++ fundus_image_toolbox/quality_prediction/models/2024-05-03 14-25-19/resnet101_summary.txt
@@ -1,20 +1,0 @@
-2024-05-03 15-04-23
-model_type: resnet101
-epochs: 25
-img_size: 350
-lr: 0.0001
-weight_decay: 0.0
-device: cuda:0
-seed: 12345
-balance_datasets: True
-use_datasets: ['drimdb', 'deepdrid-isbi2020']
-batch_size: 8
-
-Best epoch: 11
-Best val loss: 0.5173883979119377
-Best val acc: 0.7633928571428571
-test performance:
-acc: 0.7788
-auroc: 0.7779
-auprc: 0.7053
-
