+++ fundus_image_toolbox/quality_prediction/models/2024-05-03 15-52-56/efficientnet-b1_summary.txt
@@ -1,20 +1,0 @@
-2024-05-03 16-13-48
-model_type: efficientnet-b1
-epochs: 25
-img_size: 350
-lr: 0.0001
-weight_decay: 0.1
-device: cuda:0
-seed: 12345
-balance_datasets: True
-use_datasets: ['drimdb', 'deepdrid-isbi2020']
-batch_size: 8
-
-Best epoch: 21
-Best val loss: 0.5278553422706734
-Best val acc: 0.7477678571428571
-test performance:
-acc: 0.7765
-auroc: 0.7771
-auprc: 0.6979
-
