+++ fundus_image_toolbox/quality_prediction/models/2024-05-03 14-25-42/resnet50_summary.txt
@@ -1,20 +1,0 @@
-2024-05-03 14-56-44
-model_type: resnet50
-epochs: 25
-img_size: 350
-lr: 0.0001
-weight_decay: 0.0
-device: cuda:1
-seed: 12345
-balance_datasets: True
-use_datasets: ['drimdb', 'deepdrid-isbi2020']
-batch_size: 8
-
-Best epoch: 15
-Best val loss: 0.5151016148873688
-Best val acc: 0.7700892857142857
-test performance:
-acc: 0.7765
-auroc: 0.7780
-auprc: 0.6955
-

[…]
