### 2024-11-25 dc031a0
clean code and change weights path of fovea model to windows compatible string. Prevent usage of "rm" in vessel seg clone.py
