+++ fundus_image_toolbox/fovea_od_localization/model_multi.py
@@ -40,8 +40,7 @@ class ODFoveaModel:
         self.config = config
         self.device = config.device
         self.model = self._get_model(config.model_type).to(self.device)
-        self.timestamp = datetime.now().strftime("%Y-%m-%d %H:%M.%S")
-        # self.checkpoint_path = os.path.join(MODELS_DIR, f'{self.timestamp}/multi_{self.config.model_type}_best.pt')
+        self.timestamp = datetime.now().strftime("%Y-%m-%d %H_%M.%S")
         self.checkpoint_path = (
             MODELS_DIR / self.timestamp / f"multi_{self.config.model_type}_best.pt"
         )
@@ -52,9 +51,6 @@ class ODFoveaModel:
         self.optimizer = torch.optim.Adam(self.model.parameters(), lr=self.config.lr)
 
     def train(self, train_dataloader, val_dataloader):
-        # os.makedirs(os.path.dirname(self.checkpoint_path), exist_ok=True)
-        # if not os.path.exists(os.path.join(MODELS_DIR, self.timestamp, "config.yaml")):
-        #     yaml.safe_dump(vars(self.config), open(os.path.join(MODELS_DIR, self.timestamp, "config.yaml"), "w"))
         self.checkpoint_path.parent.mkdir(parents=True, exist_ok=True)
         if not (self.checkpoint_path.parent / "config.yaml").exists():
             yaml.safe_dump(
@@ -105,9 +101,7 @@ class ODFoveaModel:
         print(f"Model saved at {self.checkpoint_path}")
 
         # Save summary to file
-        # p = os.path.dirname(self.checkpoint_path)
         p = self.checkpoint_path.parent
-        # with open(os.path.join(p, "summary.txt"), 'w') as f:
         with open(p / "summary.txt", "w") as f:
             f.write(f"Best epoch: {best_epoch}\n")
             f.write(f"Best loss: {self.best_loss:.4}\n")
@@ -115,22 +109,16 @@ class ODFoveaModel:
             f.write(f'Best distance: {self.dist_tracking["val"][best_epoch]:.4}\n')
 
         # Save loss, IoU and distance tracking to files
-        # with open(os.path.join(p, f"multi_{self.config.model_type}_train_loss.txt"), 'w') as f:
         with open(p / f"multi_{self.config.model_type}_train_loss.txt", "w") as f:
             f.write("\n".join(map(str, self.loss_tracking["train"])))
-        # with open(os.path.join(p, f"multi_{self.config.model_type}_val_loss.txt"), 'w') as f:
         with open(p / f"multi_{self.config.model_type}_val_loss.txt", "w") as f:
             f.write("\n".join(map(str, self.loss_tracking["val"])))
-        # with open(os.path.join(p, f"multi_{self.config.model_type}_train_iou.txt"), 'w') as f:
         with open(p / f"multi_{self.config.model_type}_train_iou.txt", "w") as f:
             f.write("\n".join(map(str, self.iou_tracking["train"])))
-        # with open(os.path.join(p, f"multi_{self.config.model_type}_val_iou.txt"), 'w') as f:
         with open(p / f"multi_{self.config.model_type}_val_iou.txt", "w") as f:
             f.write("\n".join(map(str, self.iou_tracking["val"])))
-        # with open(os.path.join(p, f"multi_{self.config.model_type}_train_dist.txt"), 'w') as f:
         with open(p / f"multi_{self.config.model_type}_train_dist.txt", "w") as f:
             f.write("\n".join(map(str, self.dist_tracking["train"])))
-        # with open(os.path.join(p, f"multi_{self.config.model_type}_val_dist.txt"), 'w') as f:
         with open(p / f"multi_{self.config.model_type}_val_dist.txt", "w") as f:
[…]
