+++ setup.py
@@ -1,10 +1,5 @@
 from setuptools import setup, find_packages
 
-packages = find_packages()
-# packages = list(set([package.split(".")[0] for package in packages]))
-
-print(packages)
-
 setup(
     name="fundus_image_toolbox",
     version="0.0.1",
@@ -13,7 +8,7 @@ setup(
     description="A toolbox for fundus image processing",
     long_description=open("Readme.md").read(),
     long_description_content_type="text/markdown",
-    packages=packages,
+    packages=find_packages(),
     install_requires=open("requirements.txt").read().splitlines(),
     python_requires=">=3.9.0, <3.10",
     include_package_data=True,
