+++ todos.md
@@ -18,3 +18,5 @@
 [ ] Once Sarah has refactored her image cropping code to yield an image as output, add it to circle_crop, s.t. one can choose between the two algorithms
 
 [ ] Once has DOI, add citation to single and main Readmes
+
+[ ] Regularly check for TODO entries in any files
