### 2024-11-25 0e6dd74
adjust remove call for Windows
