### 2024-12-02 #7101 @osorensen
@ojasiiitd, thanks for getting started! You can generate the software paper here by writing `editorialbot generate pdf`. I'll do this for you in the next post.
