### 2024-12-02 #7101 @juliusge
Hi everyone. As suggested, we already started to address the first reviewer's points. For now,
- we fixed path issues on Windows
- we prepared for a PyPI release. For that, we changed the package structure from independent packages and a wrapper to one fully contained package. Note that this renders the manuscript's sentence "All tools can be installed as a whole or individually, depending on the user's needs." wrong and we plan to remove that sentence.
- these changes are reflected in the main branch, while the original state is in "main_before_rebuttal". As a the reviewer started to working with the main branch (thanks for joining!), we will not alter its state from now on.
