### 2024-12-02 8b32935
change: pyproject.toml for pypi release; update README; add: progress bar to weight downloading
