### 2025-01-27 7a59280
bugfix: prevent error in torch.stack
