+++ pyproject.toml
@@ -15,7 +15,7 @@ authors = [
 ]
 requires-python = ">=3.9, <3.10"
 dependencies = [
-    "bunch>=1.0.1",
+    # "bunch>=1.0.1",
     "captum>=0.7.0",
     "gdown",
     "imgaug>=0.4.0",
