### 2025-01-28 bae01ff
upd notebooks, rm highest level init
