### 2025-01-29 580c416
rename utilities to utils, present all but utils to highest level, upd notebooks and readme, reintroduce bunch dependency (needed for segmentation code), rm upper numpy version req
