### 2025-01-29 0fd5aea
add dev dependency, del outdated files
