+++ outdated/fundus_utilities__/test/test_image_torch_utils.py
@@ -1,139 +1,0 @@
-import unittest
-import matplotlib.pyplot as plt
-import numpy as np
-from PIL import Image
-import cv2
-import torch
-from torchvision.transforms.functional import to_tensor
-from fundus_utilities import ImageTorchUtils as Img
-
-fundus1_path = 'fundus1.jpg'
-fundus2_path = 'fundus2.jpg'
-
-fundus1_plt = plt.imread(fundus1_path)
-fundus1_img = Image.open(fundus1_path)
-fundus1_cv2 = cv2.imread(fundus1_path)
-
-class TestImageTorchUtils(unittest.TestCase):
-    def _to_tensor(self):
-        # Test the to_tensor function
-        imgs_rgb = [fundus1_path, fundus1_plt, fundus1_img]
-
-        # Convert to rgb tensor from rgb image
-        outs = [Img(img).to_tensor().img for img in imgs_rgb]
-
-        # Convert to rgb tensor from bgr image
-        bgr_out = Img(fundus1_cv2).to_tensor(from_cspace='bgr').img
-        # Alternatively, use a sequence of calls:
-        bgr_out = Img(fundus1_cv2).to_tensor().to_cspace(from_cspace='bgr', to_cspace='rgb').img
-        outs.append(bgr_out)
-
-        assert all([torch.equal(outs[0], out) for out in outs[1:]])
-
-    def _to_cspace(self):
-        # Test the to_cspace function
-        imgs_rgb = [fundus1_path, fundus1_plt, fundus1_img]
-        img_bgr = fundus1_cv2
-
-        # Convert from rgb to bgr
-        outs = [Img(img).to_tensor().to_cspace(from_cspace='rgb', to_cspace='bgr').img for img in imgs_rgb]
-
-        assert all([torch.equal(outs[0], out) for out in outs[1:]])
-        assert torch.equal(Img(img_bgr).to_tensor().img, outs[0])
-
-        # Convert to grey scale
-        outs = [Img(img).to_tensor().to_cspace(from_cspace='rgb', to_cspace='grey').img for img in imgs_rgb]
-
-        assert all([torch.equal(outs[0], out) for out in outs[1:]])
-        assert torch.equal(Img(img_bgr).to_tensor().to_cspace(from_cspace='bgr', to_cspace='grey').img, outs[0])
-
-    def _set_channel_dim(self):
-        # Test set_channel_dim function
-        imgs = [fundus1_path, fundus1_plt, fundus1_img, fundus1_cv2]
-
-        # Set channel dimension to the last dimension
-        outs = [Img(img).to_tensor().set_channel_dim(-1).img for img in imgs]
-
-        # Check if the channel dimension is the last dimension
-        assert all([out.shape[-1] == 3 for out in outs])
-    
-    def _to_pil(self):
-        # Test to_pil function
-        imgs = [fundus1_path, fundus1_plt, fundus1_img, fundus1_cv2]
-
-        # Convert to pil image
-        outs = [Img(img).to_tensor().to_pil().img for img in imgs]
-
-        # Check if the output is a pil image
-        assert all([isinstance(out, Image.Image) for out in outs])
-
-    def _to_numpy(self):
-        # Test to_numpy functions
-        imgs = [fundus1_path, fundus1_plt, fundus1_img, fundus1_cv2]
-
-        # Convert to numpy array
-        outs = [Img(img).to_tensor().to_numpy().img for img in imgs]
-
-        # Check if the output is a numpy array
-        assert all([isinstance(out, np.ndarray) for out in outs])
-        assert all([out.dtype == np.uint8 for out in outs])
-
-        # Alternatively, chain the to_numpy and to_unit8 functions
-        outs = [Img(img).to_tensor().to_numpy().to_uint8().img for img in imgs]
-
[…]
