+++ outdated/fundus_registration__/requirements.txt
@@ -1,19 +1,0 @@
-ipykernel>=6.15.2
-ipython>=8.16.1
-jupyter>=1.0.0
-jupyter-black>=0.3.4
-imgaug>=0.4.0
-matplotlib>=3.6.2
-numpy>=1.24.0
-opencv-python-headless>=4.9.0.80
-pillow>=10.2.0
-PyYAML>=6.0
-scikit-learn>=1.1.2
-scipy>=1.10.0
-torch>=2.2.0
-torchvision>=0.15.2
-tqdm>=4.66.2
-ipykernel>=6.15.2
-ipython>=8.16.1
-gdown
-fundus_utilities @ git+https://github.com/berenslab/fundus_image_toolbox#egg=fundus_utilities&subdirectory=fundus_utilities

[…]
