+++ fundus_image_toolbox/vessel_segmentation/clone.py
@@ -103,3 +103,13 @@ def add_dots_to_imports_in_folder(folder_path):
             if file.endswith(".py"):
                 print("########\n", file)
                 add_dots_to_imports(os.path.join(root, file))
+
+def replace_timm(folder_path):
+    for root, _, files in os.walk(folder_path):
+        for file in files:
+            if file.endswith(".py"):
+                with open(os.path.join(root, file), "r") as f:
+                    lines = f.readlines()
+                with open(os.path.join(root, file), "w") as f:
+                    for line in lines:
+                        f.write(line.replace("timm.models.layers", "timm.layers"))
