### 2025-01-29 6b650f3
relax requirements in pyproject.toml
