+++ pyproject.toml
@@ -15,7 +15,7 @@ authors = [
 ]
 requires-python = ">=3.9"
 dependencies = [
-    "albumentations>=2.0.2",
+    "albumentations",
     "bunch",
     "captum>=0.7.0",
     "gdown>=5.2.0",
@@ -24,27 +24,27 @@ dependencies = [
     "ipython>=8.18.1",
     "jupyter>=1.1.1",
     "jupyter-black>=0.4.0",
-    "matplotlib>=3.9.4",
-    "matplotlib-inline>=0.1.7",
+    "matplotlib>=3.6.2",
+    "matplotlib-inline>=0.1.6",
     "munch>=4.0.0",
-    "mypy>=1.14.1",
+    "mypy>=0.910",
     "notebook-shim>=0.2.4",
-    "numpy>=2.0.2",
-    "opencv-python-headless>=4.11.0.86",
-    "openpyxl>=3.1.5",
-    "pandas>=2.2.3",
-    "pillow>=11.1.0",
+    "numpy>=1.24.0",
+    "opencv-python-headless>=4.9.0.80",
+    "openpyxl>=3.1.2",
+    "pandas>=1.4.2",
+    "pillow>=10.2.0",
     "pip>=25.0",
-    "pyyaml>=6.0.2",
-    "scikit-image>=0.24.0",
-    "scikit-learn>=1.6.1",
-    "scipy>=1.13.1",
+    "pyyaml>=6.0.0",
+    "scikit-image>=0.22.0",
+    "scikit-learn>=1.1.2",
+    "scipy>=1.10.0",
     "seaborn>=0.13.2",
-    "timm>=1.0.14",
-    "torch>=2.5.1",
+    "timm>=0.9.16",
+    "torch>=2.2.0",
     "torchinfo>=1.8.0",
-    "torchvision>=0.20.1",
-    "tqdm>=4.67.1",
+    "torchvision>=0.25.2",
+    "tqdm>=4.66.2",
 ]
 
 [dependency-groups]
