+++ pyproject.toml
@@ -43,7 +43,7 @@ dependencies = [
     "timm>=0.9.16",
     "torch>=2.2.0",
     "torchinfo>=1.8.0",
-    "torchvision>=0.25.2",
+    "torchvision>=0.15.2",
     "tqdm>=4.66.2",
 ]
 
