### 2025-01-31 cc0d486
replace os with pathlib for Windows compatibility
