### 2025-01-31 852a6dd
specify weights_only to avoid deprecation warning
