+++ fundus_image_toolbox/registration/inference.py
@@ -210,7 +210,7 @@ def load_model(config: dict = None):
 
     model = SuperRetina().to(device)
     model.load_state_dict(
-        torch.load(config["model_save_path"], map_location=device)["net"]
+        torch.load(config["model_save_path"], map_location=device)["net"], weights_only=False
     )
     model.eval()
 
