### 2025-01-31 3fbaa7d
Update paper.md
Add Ziwei, mention PyPI, remove independence of submodules statement, as it is no longer correct
