+++ 1_read_more/Readmore_quality_prediction.md
@@ -1,7 +1,7 @@
 ## A quality prediction model for fundus images (gradeable vs. ungradeable)
 Based on a 10-model ensemble (ResNets and EfficientNets) trained on DeepDRiD and DrimDB data. For DrimDB images, we considered out-of-domain images such as photographs of the outer eye as ungradeable.
 
-<img src="../quality_prediction/ex.png" alt="Example image" width="800"/>
+<img src="../fundus_image_toolbox/quality_prediction/ex.png" alt="Example image" width="800"/>
 
 <br>
 [x] Works on tensor images <br>
@@ -31,4 +31,6 @@ Based on a 10-model ensemble (ResNets and EfficientNets) trained on DeepDRiD and
 - The model ensemble achieves an accuracy and ROCAUC of 0.78 and 0.84 on a DeepDRiD test split and 1.0 and 1.0 on a DrimDB test split.
 - To predict DeepDRiD's binary quality annotation is known to be a difficult task. Hence, the previous sota is 0.75 and 0.75 for accuracy and ROCAUC, respectively, on a DeepDRiD test split ([Tummala et al., 2023](https://doi.org/10.3390/diagnostics13040622)).
 
+
 ### Cite
+See [Readme.md](../Readme.md#cite).
