### 2025-02-03 51441b9
Replace from source dependency bunch with on-the-fly git clone for PyPI
