+++ pyproject.toml
@@ -9,7 +9,7 @@ write_to = "fundus_image_toolbox/_version.py"
 name = "fundus_image_toolbox"
 version = "0.0.4"
 description = "A toolbox for fundus image analysis"
-readme = {file = "README.md", content-type = "text/markdown"}
+readme = {file = "Readme.md", content-type = "text/markdown"}
 authors = [
     {name = "Julius Gervelmeyer", email = "Julius.Gervelmeyer@uni-tuebingen.de"},
 ]
@@ -59,3 +59,6 @@ where = ["."]
 
 [tool.uv.workspace]
 members = ["."]
+
+[project.urls]
+repository = "https://github.com/berenslab/fundus_image_toolbox"
