### 2025-02-03 #7101 @juliusge
We are set now: We addressed the reviewer's comments (saeedahmadicp) in the main branch. Specifically, we:
- updated and extended the usage examples.
- fixed compatibility issues with Windows.
- added contribution guidelines to the Readme.md.
- added author contributions to the paper.
- removed from the paper the statement about the submodules being standalone packages. We had to remove that feature in order to meet the PyPI requirements.
- reworked the code s.t. it is compatible with PyPI releases and added v0.0.4 to PyPI.
