### 2025-02-11 8ded196
minor upd of paper.md
