### 2025-03-18 #7101 @juliusge
Thanks everyone!
- The new tagged version is v0.1.0
- The archive DOI is https://doi.org/10.5281/zenodo.15046293
