### 2025-03-26 #7101 @osorensen
> [@editorialbot](https://github.com/editorialbot) generate pdf
>
> 🔍 checking out the following:
>
> * [ ]  reviewer checklists are completed or addressed[ ]  version set[ ]  archive set[ ]  archive names (including order) and title in archive matches those specified in the paper[ ]  archive uses the same license as the repo and is OSI approved as open source[ ]  archive DOI and version match or redirect to those set by editor in review thread[ ]  paper is error free - grammar and typos[ ]  paper is error free - test links in the paper and bib[ ]  paper is error free - refs preserve capitalization where necessary[ ]  paper is error free - no invalid refs without justification

@crvernon, I haven't seen this in papers I've edited before. Am I supposed to go through the checklist, or is it the editor-in-chief?
