### 2025-04-03 #7101 @arfon
> @crvernon, I haven't seen this checklist in papers I've edited before. Am I supposed to go through the checklist, or is it the editor-in-chief?

This is Chris' list for himself :-)
